space_deploy/requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
openenv-core>=0.2.3
huggingface_hub>=0.26.0
hf_transfer>=0.1.8
transformers>=4.45.0
datasets>=3.0.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.12.0
torch>=2.4.0
torchvision>=0.19.0
bitsandbytes>=0.43.0
sentencepiece>=0.2.0
protobuf>=4.25.0
xformers>=0.0.27
python-dotenv>=1.0.0
